requirements.txt
absl-py==1.3.0
alphashape==1.3.1
anyio==3.5.0
anytree==2.8.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-log==0.4.0
colorama==0.4.4
ConfigArgParse==1.5.3
contextlib2==21.6.0
coverage==6.4.4
cycler==0.11.0
Cython==0.29.30
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
einops==0.6.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
flake8==4.0.1
Flask==2.2.3
fonttools==4.33.3
idna==3.3
importlib-metadata==6.0.0
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jakteristics==0.5.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.3.1
jupyter-server==1.23.6
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.2
laspy==2.1.2
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
ml-collections==0.1.1
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.7.0
nest-asyncio==1.5.5
networkx==2.8.2
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.23.1
open3d==0.17.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
pip-licenses==3.5.4
plotly==5.13.1
prometheus-client==0.13.1
prompt-toolkit==3.0.29
psutil==5.9.0
PTable==0.9.2
PuLP==2.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
PyMaxflow==1.2.13
pyparsing==3.0.7
pyproj==3.3.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
Rtree==1.0.0
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
tenacity==8.2.2
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
toml==0.10.2
tomlkit==0.10.1
tornado==6.1
traitlets==5.1.1
trimesh==3.12.4
typer==0.4.1
typing_extensions==4.4.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.2.3
widgetsnbextension==4.0.5
zipp==3.15.0
